Changes the way the whole system works. Prepares it for the automatic evolution of classifiers when a segment comes into the scene.
Adds the "Invoker" abstract class as well as the "Visualizer" and the
corrected "VideoFeeder" interfaces.
Makes the whole system "Invoker" centric: all communications must go
through the "Invoker" object.
Makes the VideoFeeder passive: it no longer automatically passes frames.
The "Invoker" asks for the next frame.

Current bug: when a classifier is disabled in the visualizer, since it
is not seen by the system, a new classifier is evolved to replace it!